+++ app/controllers/sessions_controller.rb
@@ -1,5 +1,4 @@
 class SessionsController < ApplicationController
   include Godmin::Authentication::SessionsController
 
-  binding.pry
 end
