+++ config/routes.rb
@@ -1,6 +1,5 @@
 Rails.application.routes.draw do
 
-
   namespace :api do
     namespace :v1 do
       resources :users, only: [:index, :show], constraints: {format: /(text|json)/}
@@ -12,9 +11,9 @@ Rails.application.routes.draw do
   end
 
   devise_for :users,
-             path: 'api/v1/users',
-             controllers: {registrations: 'api/v1/registrations', sessions: 'api/v1/sessions'},
-             skip: [:sessions]
+              path: 'api/v1/users',
+              controllers: {registrations: 'api/v1/registrations'},
+              skip: [:sessions]
   resources :users
   resource :session, only: [:new, :create, :destroy]
   root to: 'application#welcome'
