+++ spec/factories/users.rb
@@ -1,6 +1,7 @@
 FactoryGirl.define do
   factory :user do
     user_name 'Random Guy'
-    email { Faker.email}
+    email { Faker::Internet.email }
+    password { Faker::Internet.password(8) }
   end
 end
